🐛 fix : rename "services" folder to "utils"
