+++ src/components/input/index.tsx
@@ -1,5 +1,5 @@
 import { DetailedHTMLProps, InputHTMLAttributes, FunctionComponent } from 'react';
-import { mergeClassNames } from '../../services/classes';
+import { mergeClassNames } from '../../utils/classes';
 import './styles.scss';
 
 export interface InputProps
